+++ examples/stream_blocks_single.rs
@@ -6,7 +6,7 @@ use std::env;
 use std::pin::pin;
 
 use geyser_grpc_connector::grpc_subscription_autoreconnect_streams::{
-    create_geyser_reconnecting_stream, GeyserFilter,
+    create_geyser_reconnecting_stream,
 };
 use geyser_grpc_connector::grpcmultiplex_fastestwins::{
     create_multiplexed_stream, FromYellowstoneExtractor,
@@ -16,8 +16,8 @@ use tracing::warn;
 use yellowstone_grpc_proto::geyser::subscribe_update::UpdateOneof;
 use yellowstone_grpc_proto::geyser::SubscribeUpdate;
 use yellowstone_grpc_proto::prost::Message as _;
-use geyser_grpc_connector::grpc_subscription_autoreconnect_tasks::{create_geyser_reconnecting_task, Message};
-use geyser_grpc_connector::{GrpcConnectionTimeouts, GrpcSourceConfig};
+use geyser_grpc_connector::grpc_subscription_autoreconnect_tasks::{create_geyser_autoconnection_task, Message};
+use geyser_grpc_connector::{GeyserFilter, GrpcConnectionTimeouts, GrpcSourceConfig};
 
 fn start_example_blockmini_consumer(
     multiplex_stream: impl Stream<Item = BlockMini> + Send + 'static,
@@ -96,13 +96,13 @@ pub async fn main() {
 
     info!("Write Block stream..");
 
-    let (jh_geyser_task, mut green_stream) = create_geyser_reconnecting_task(
+    let (jh_geyser_task, mut green_stream) = create_geyser_autoconnection_task(
         green_config.clone(),
         GeyserFilter(CommitmentConfig::confirmed()).blocks_and_txs(),
     );
 
     tokio::spawn(async move {
-        while let Ok(message) = green_stream.recv().await {
+        while let Some(message) = green_stream.recv().await {
             match message {
                 Message::GeyserSubscribeUpdate(subscriber_update) => {
                     // info!("got update: {:?}", subscriber_update.update_oneof.);
