+++ src/lib.rs
@@ -1,5 +1,8 @@
+use std::collections::HashMap;
 use std::fmt::{Debug, Display};
 use std::time::Duration;
+use solana_sdk::commitment_config::CommitmentConfig;
+use yellowstone_grpc_proto::geyser::{CommitmentLevel, SubscribeRequest, SubscribeRequestFilterBlocks, SubscribeRequestFilterBlocksMeta};
 use yellowstone_grpc_proto::tonic::transport::ClientTlsConfig;
 
 pub mod grpc_subscription;
@@ -61,6 +64,75 @@ impl GrpcSourceConfig {
             grpc_x_token,
             tls_config,
             timeouts: Some(timeouts),
+        }
+    }
+}
+
+
+#[derive(Clone)]
+pub struct GeyserFilter(pub CommitmentConfig);
+
+impl GeyserFilter {
+    pub fn blocks_and_txs(&self) -> SubscribeRequest {
+        let mut blocks_subs = HashMap::new();
+        blocks_subs.insert(
+            "client".to_string(),
+            SubscribeRequestFilterBlocks {
+                account_include: Default::default(),
+                include_transactions: Some(true),
+                include_accounts: Some(false),
+                include_entries: Some(false),
+            },
+        );
+
+        SubscribeRequest {
+            slots: HashMap::new(),
+            accounts: Default::default(),
+            transactions: HashMap::new(),
+            entry: Default::default(),
+            blocks: blocks_subs,
+            blocks_meta: HashMap::new(),
+            commitment: Some(map_commitment_level(self.0) as i32),
+            accounts_data_slice: Default::default(),
+            ping: None,
+        }
+    }
+
+    pub fn blocks_meta(&self) -> SubscribeRequest {
+        let mut blocksmeta_subs = HashMap::new();
+        blocksmeta_subs.insert("client".to_string(), SubscribeRequestFilterBlocksMeta {});
+
+        SubscribeRequest {
+            slots: HashMap::new(),
+            accounts: Default::default(),
+            transactions: HashMap::new(),
+            entry: Default::default(),
+            blocks: HashMap::new(),
+            blocks_meta: blocksmeta_subs,
+            commitment: Some(map_commitment_level(self.0) as i32),
+            accounts_data_slice: Default::default(),
+            ping: None,
+        }
+    }
+}
+
+fn map_commitment_level(commitment_config: CommitmentConfig) -> CommitmentLevel {
+    // solana_sdk -> yellowstone
+    match commitment_config.commitment {
+        solana_sdk::commitment_config::CommitmentLevel::Processed => {
+            yellowstone_grpc_proto::prelude::CommitmentLevel::Processed
+        }
+        solana_sdk::commitment_config::CommitmentLevel::Confirmed => {
+            yellowstone_grpc_proto::prelude::CommitmentLevel::Confirmed
+        }
+        solana_sdk::commitment_config::CommitmentLevel::Finalized => {
+            yellowstone_grpc_proto::prelude::CommitmentLevel::Finalized
+        }
+        _ => {
+            panic!(
+                "unsupported commitment level {}",
+                commitment_config.commitment
+            )
         }
     }
 }
